Yjs interagtion (not quite all functions)
